+++ src/main/java/eu/europa/ted/eforms/viewer/config/NoticeViewerConfig.java
@@ -5,6 +5,9 @@ import java.io.OutputStream;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.StandardOpenOption;
+import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.List;
 import java.util.Optional;
 import org.apache.commons.io.IOUtils;
 import org.apache.commons.lang3.Validate;
@@ -26,14 +29,14 @@ public class NoticeViewerConfig {
   private NoticeViewerConfig() {}
 
   /**
-   * Create a Freemarker configuration for locating templates for loading.
-   * The templates are first looked up on the filesystem.
-   * If not found, then they are loaded from the templates folder bundled with the application.
-   * The templates root directory can be configured using the system property {@value NoticeViewerConstants#TEMPLATES_ROOT_DIR_PROPERTY}
+   * Create a Freemarker configuration for locating templates for loading. The templates are first
+   * looked up on the filesystem. If not found, then they are loaded from the templates folder
+   * bundled with the application. The templates root directory can be configured using the system
+   * property {@value NoticeViewerConstants#TEMPLATES_ROOT_DIR_PROPERTY}
    *
    * @return A Freemarker configuration object.
-   * @throws IOException 
-   * @throws Exception 
+   * @throws IOException
+   * @throws Exception
    */
   public static Configuration getFreemarkerConfig() throws IOException {
     if (freemarkerConfig == null) {
@@ -45,14 +48,19 @@ public class NoticeViewerConfig {
       logger.debug("Configuring Freemarker using [{}] as the templates root directory.",
           templatesRootDirPath);
 
-      populateExternalTemplatesDir(templatesRootDirPath);
+      List<TemplateLoader> templateLoaders = new ArrayList<>(
+          Arrays.asList(new ClassTemplateLoader(NoticeViewerConfig.class, "/templates")));
 
-      FileTemplateLoader fileTemplateLoader =
-          new FileTemplateLoader(templatesRootDirPath.toFile());
-      ClassTemplateLoader classTemplateLoader =
-          new ClassTemplateLoader(NoticeViewerConfig.class, "/templates");
+      try {
+        populateExternalTemplatesDir(templatesRootDirPath);
+        templateLoaders.add(new FileTemplateLoader(templatesRootDirPath.toFile()));
+      } catch (IOException e) {
+        logger.warn("Failed to populate external templates directory [{}]", templatesRootDirPath);
+        logger.debug("The error was:", e);
+      }
+
       MultiTemplateLoader multiTemplateLoader =
-          new MultiTemplateLoader(new TemplateLoader[] {fileTemplateLoader, classTemplateLoader});
+          new MultiTemplateLoader(templateLoaders.toArray(TemplateLoader[]::new));
 
       freemarkerConfig = new Configuration(Configuration.VERSION_2_3_31);
       freemarkerConfig.setTemplateLoader(multiTemplateLoader);
@@ -64,12 +72,12 @@ public class NoticeViewerConfig {
   }
 
   /**
-   * Populates a templates folder (external to the application) with the bundled templates.
-   * Any existing templates on the target folder will not be overwritten.
+   * Populates a templates folder (external to the application) with the bundled templates. Any
+   * existing templates on the target folder will not be overwritten.
    *
    * @param targetTemplatesRootDir The target templates root directory on the filesystem
-   * @throws IOException 
-   * @throws Exception 
+   * @throws IOException
+   * @throws Exception
    */
   private static void populateExternalTemplatesDir(Path targetTemplatesRootDir) throws IOException {
     Validate.notNull(targetTemplatesRootDir, "Undefined templates root directory");
@@ -84,7 +92,9 @@ public class NoticeViewerConfig {
         logger.debug("Copying template [classpath:{}] to [{}]", sourcePath,
             targetPath.toAbsolutePath());
 
-        Files.createDirectories(targetPath.getParent());
+        Path parentPath = targetPath.getParent();
+        Files.createDirectories(parentPath);
+
         try (OutputStream os = Files.newOutputStream(targetPath, StandardOpenOption.CREATE)) {
           IOUtils.copy(cl.getResourceAsStream(sourcePath.toString().replace("\\", "/")), os);
         }