+++ CHANGELOG.md
@@ -1,11 +1,12 @@
-# eForms Notice Viewer 0.6.0 - Release Notes
+# eForms Notice Viewer 0.7.0 - Release Notes
 
 _The eForms Notice Viewer is a sample command line application that demonstrates how you can use the [eForms SDK](https://github.com/OP-TED/eForms-SDK) in a metadata driven application that visualises eForms notices._
 
 ---
 ## In this release:
 
-* We moved some reusable classes from Notice Viewer to the EFX Toolkit and eForms Core libraries.
+* Introduced usage of Freemarker templates for the generation of XSL.
+* Implemented caching of the generated XSL.
 
 --- 
 
@@ -13,4 +14,4 @@ Documentation for this sample application is available at: https://docs.ted.euro
 
 This version depends on:
  - [EFX toolkit for Java](https://github.com/OP-TED/efx-toolkit-java) version 1.2.0.
- - [eForms Core for Java](https://github.com/OP-TED/eforms-core-java) library version 1.0.0.
+ - [eForms Core for Java](https://github.com/OP-TED/eforms-core-java) library version 1.0.1.